Implemented Coding Standard

+++ Profile_App/form/delete.php
@@ -1,15 +1,24 @@
 <?php
-require_once("config/mysql_connect_db.php");
-if (isset($_GET['emp_id'])) {
+/**
+ * Delete employee data file.
+ */
+
+require_once('config/database_config.php');
+
+if (isset($_GET['emp_id']) && preg_match('/^[0-9]*$/',$_GET['emp_id'])) {
     $emp_id = stripslashes($_GET['emp_id']);
     $photo = 'profile_pic/'.stripslashes($_GET['photo']);
+
+    // Delete data from address table
     $sql='DELETE FROM address WHERE employee_id ='. $emp_id;
     mysqli_query($conn,$sql);
+
+    // Delete data from employee table
 	$sql='DELETE FROM employee WHERE id ='.$emp_id;
 	mysqli_query($conn,$sql);
-	unlink($photo);
-       
+
+    // Delete profile pic
+    unlink($photo);
 }
-header("Location: output.php");
-    
-?>
+
+header('Location: output.php');
